BigSegmentHeader: system property to specify LOCK_TIMEOUT_SECONDS
The default of 60 seconds may be too long if using Chronicle-Map in a
way that prevents real interprocess dead locks, i.e.
InterProcessDeadLockException can only be raised if a previous
Chronicle-Map process has been killed while holding a lock. In this
case, a timeout of e.g 5 seconds (or less) is sufficient to detect the
problem and possibly decide to invoke a recovery.